Updating to latest

+++ frontend/src/app/login/login.component.ts
@@ -5,10 +5,11 @@ import { Component, OnInit } from '@angular/core'
 import { FormControl, Validators } from '@angular/forms'
 import { library, dom } from '@fortawesome/fontawesome-svg-core'
 import { UserService } from '../Services/user.service'
-import { faKey } from '@fortawesome/free-solid-svg-icons'
+import { faKey, faEye, faEyeSlash } from '@fortawesome/free-solid-svg-icons'
 import { faGoogle } from '@fortawesome/free-brands-svg-icons'
+import { FormSubmitService } from '../Services/form-submit.service'
 
-library.add(faKey, faGoogle)
+library.add(faKey, faEye, faEyeSlash, faGoogle)
 dom.watch()
 
 const oauthProviderUrl = 'https://accounts.google.com/o/oauth2/v2/auth'
@@ -25,9 +26,7 @@ const authorizedRedirectURIs = {
   'http://127.0.0.1:3000': 'http://local3000.owasp-juice.shop',
   'http://localhost:4200': 'http://local4200.owasp-juice.shop',
   'http://127.0.0.1:4200': 'http://local4200.owasp-juice.shop',
-  'http://192.168.99.100:3000': 'http://localMac.owasp-juice.shop',
-  'https://juice-shop-v8.herokuapp.com': 'https://juice-shop-v8.herokuapp.com',
-  'http://juice-shop-v8.herokuapp.com': 'http://juice-shop-v8.herokuapp.com'
+  'http://192.168.99.100:3000': 'http://localMac.owasp-juice.shop'
 }
 
 @Component({
@@ -37,15 +36,15 @@ const authorizedRedirectURIs = {
 })
 export class LoginComponent implements OnInit {
 
-  public emailControl = new FormControl('', [ Validators.required])
-  public passwordControl = new FormControl('', [ Validators.required])
+  public emailControl = new FormControl('', [Validators.required])
+  public passwordControl = new FormControl('', [Validators.required])
   public hide = true
   public user: any
   public rememberMe: FormControl = new FormControl(false)
   public error: any
   public oauthUnavailable: any
   public redirectUri
-  constructor (private userService: UserService, private windowRefService: WindowRefService, private cookieService: CookieService, private router: Router) { }
+  constructor (private userService: UserService, private windowRefService: WindowRefService, private cookieService: CookieService, private router: Router, private formSubmitService: FormSubmitService) { }
 
   ngOnInit () {
 
@@ -63,6 +62,8 @@ export class LoginComponent implements OnInit {
     if (this.oauthUnavailable) {
       console.log(this.redirectUri + ' is not an authorized redirect URI for this application.')
     }
+
+    this.formSubmitService.attachEnterKeyHandler('login-form', 'loginButton', () => this.login())
   }
 
   login () {
@@ -73,13 +74,20 @@ export class LoginComponent implements OnInit {
     this.userService.login(this.user).subscribe((authentication: any) => {
       localStorage.setItem('token', authentication.token)
       this.cookieService.put('token', authentication.token)
-      sessionStorage.setItem('bid',authentication.bid)
+      sessionStorage.setItem('bid', authentication.bid)
       /*Use userService to notifiy if user has logged in*/
       /*this.userService.isLoggedIn = true;*/
       this.userService.isLoggedIn.next(true)
       this.router.navigate(['/search'])
-    }, (error) => {
+    }, ({ error }) => {
+      if (error.status && error.data && error.status === 'totp_token_requried') {
+        localStorage.setItem('totp_tmp_token', error.data.tmpToken)
+        this.router.navigate(['/2fa/enter'])
+        return
+      }
+
       console.log(error)
+
       localStorage.removeItem('token')
       this.cookieService.remove('token', { domain: document.domain })
       sessionStorage.removeItem('bid')
@@ -105,8 +113,8 @@ export class LoginComponent implements OnInit {
   googleLogin () {
 
     this.windowRefService.nativeWindow.location.replace(oauthProviderUrl + '?client_id='
-    + clientId + '&response_type=token&scope=email&redirect_uri='
-    + authorizedRedirectURIs[this.redirectUri])
+      + clientId + '&response_type=token&scope=email&redirect_uri='
+      + authorizedRedirectURIs[this.redirectUri])
 
   }
